Join Ensemble and LegacyEnsemble
